requirements.txt
PySide2
deepspeech
webrtcvad-wheels
pydub
moviepy
qtmodern